Various components improvements - move ListItemLink component to titus-kitchen-sink (to avoid having a dependency on reach/router in titus-components) - install missing prop-types dependency in titus-components - move nodemon dependency to devDependencies - update packages
