+++ packages/titus-components/src/index.d.ts
@@ -1,5 +1,4 @@
 export * from './autocomplete'
-export * from './list-item-link'
 export * from './navigation'
 export * from './table'
 export * from './uploader'
