+++ packages/titus-components/src/index.js
@@ -1,4 +1,3 @@
-export { default as ListItemLink } from './list-item-link/list-item-link'
 export { default as Navigation } from './navigation/navigation'
 export { default as Table } from './table/table'
 export { default as Autocomplete } from './autocomplete/autocomplete'
