--- packages/titus-components/src/list-item-link/index.d.ts
@@ -1,2 +1,0 @@
-export { default, default as ListItemLink } from './list-item-link'
-export * from './list-item-link'
