+++ packages/titus-kitchen-sink/package.json
@@ -6,14 +6,14 @@
     "@nearform/commentami-react-components": "^1.0.5",
     "@nearform/react-table": "^1.1.0",
     "@nearform/titus-components": "^1.0.0",
-    "@reach/router": "^1.1.1",
+    "@reach/router": "^1.2.1",
     "@vx/gradient": "0.0.165",
     "@vx/group": "0.0.170",
     "@vx/hierarchy": "0.0.170",
     "@vx/shape": "0.0.170",
     "@vx/text": "0.0.165",
-    "apollo-boost": "^0.1.10",
-    "apollo-link": "^1.2.2",
+    "apollo-boost": "^0.1.16",
+    "apollo-link": "^1.2.3",
     "auth0-js": "^9.7.3",
     "cross-env": "^5.2.0",
     "fast-levenshtein": "^2.0.6",
@@ -22,9 +22,9 @@
     "i18next": "^11.9.0",
     "i18next-browser-languagedetector": "^2.2.3",
     "i18next-xhr-backend": "^1.5.1",
-    "react": "^16.5.0",
-    "react-apollo": "^2.1.6",
-    "react-dom": "^16.5.0",
+    "react": "^16.5.2",
+    "react-apollo": "^2.2.2",
+    "react-dom": "^16.5.2",
     "react-i18next": "^7.12.0",
     "react-loadable": "^5.5.0",
     "react-scripts": "1.1.4",
@@ -32,7 +32,7 @@
     "reselect": "^3.0.1",
     "sillyname": "^0.1.0",
     "victory": "^0.27.0",
-    "yup": "^0.26.5"
+    "yup": "^0.26.6"
   },
   "scripts": {
     "start": "react-app-rewired start",
@@ -49,12 +49,12 @@
     "lighthouse": "node lighthouse"
   },
   "devDependencies": {
-    "@material-ui/core": "^3.0.1",
+    "@material-ui/core": "^3.1.1",
     "@material-ui/icons": "^3.0.1",
-    "@storybook/addon-actions": "^3.4.7",
-    "@storybook/addon-links": "^3.4.7",
-    "@storybook/addons": "^3.4.7",
-    "@storybook/react": "^3.4.7",
+    "@storybook/addon-actions": "^3.4.11",
+    "@storybook/addon-links": "^3.4.11",
+    "@storybook/addons": "^3.4.11",
+    "@storybook/react": "^3.4.11",
     "babel-core": "^6.26.3",
     "babel-eslint": "^7.2.3",
     "babel-plugin-import-graphql": "^2.6.2",
@@ -62,7 +62,7 @@
     "create-react-class": "^15.6.3",
     "eslint": "^4.1.1",
     "eslint-config-react-app": "^2.1.0",
-    "eslint-plugin-flowtype": "^2.34.1",
+    "eslint-plugin-flowtype": "^2.50.3",
     "eslint-plugin-import": "^2.6.0",
     "eslint-plugin-jsx-a11y": "^5.1.1",
     "eslint-plugin-react": "^7.1.0",
@@ -70,14 +70,14 @@
     "hapi-swagger": "^9.1.1",
     "husky": "^0.14.3",
     "inert": "^5.1.0",
-    "jest-dom": "^1.8.1",
-    "lighthouse": "^3.1.1",
+    "jest-dom": "^1.12.1",
+    "lighthouse": "^3.2.0",
     "prettier": "1.14.2",
-    "pretty-quick": "^1.6.0",
+    "pretty-quick": "^1.7.0",
     "react-app-rewire-graphql-tag": "^1.1.0",
     "react-app-rewired": "^1.6.2",
     "react-testing-library": "^4.1.2",
-    "serve": "^10.0.0",
+    "serve": "^10.0.2",
     "vision": "^5.4.0"
   },
   "proxy": "http://localhost:5000"
