+++ packages/titus-kitchen-sink/src/components/temporal/queries.graphql
@@ -6,7 +6,7 @@ query loadFoodHistoryData($id: ID!) {
 
   foodHistory(id: $id) {
     # we explicitly don't return the id to avoid apollo caching by id
-    # because id is the same or all histories of a food
+    # because id is the same for all histories of a food
     name,
     created,
     modified,
